Revert "Refactored Notes.jsx"
