--- Client/src/components/notes/NotesList.jsx
@@ -1,114 +1,0 @@
-import NoteCard from "./NoteCard";
-
-const NotesList = ({
-  selectedNote,
-  pinnedNotes,
-  unpinnedNotes,
-  filteredNotes,
-  searchTerm,
-  setSelectedNote,
-  togglePin,
-  deleteNote,
-  duplicateNote,
-  exportNote,
-  changeColor,
-  showColorPicker,
-  setShowColorPicker,
-  colors,
-  getPlainTextPreview,
-}) => {
-  return (
-    <div
-      className={`${selectedNote ? "w-80" : "w-full"} ${
-        selectedNote ? "border-r" : ""
-      } overflow-auto p-4`}
-      style={{
-        backgroundColor: "var(--bg-primary)",
-        borderColor: "var(--bg-sec)",
-      }}
-    >
-      {/* Pinned notes */}
-      {pinnedNotes.length > 0 && (
-        <div className="mb-6">
-          <h3
-            className="text-xs font-medium uppercase mb-2 mt-0"
-            style={{ color: "var(--txt-dim)" }}
-          >
-            Pinned
-          </h3>
-          <div
-            className="grid gap-2"
-            style={{
-              gridTemplateColumns: "repeat(auto-fill, minmax(200px, 1fr))",
-            }}
-          >
-            {pinnedNotes.map((note) => (
-              <NoteCard
-                key={note.id}
-                note={note}
-                onSelect={setSelectedNote}
-                onPin={togglePin}
-                onDelete={deleteNote}
-                onDuplicate={duplicateNote}
-                onExport={exportNote}
-                onColorChange={changeColor}
-                showColorPicker={showColorPicker}
-                setShowColorPicker={setShowColorPicker}
-                colors={colors}
-                getPlainTextPreview={getPlainTextPreview}
-              />
-            ))}
-          </div>
-        </div>
-      )}
-
-      {/* Unpinned notes */}
-      {unpinnedNotes.length > 0 && (
-        <div>
-          {pinnedNotes.length > 0 && (
-            <h3
-              className="text-xs font-medium uppercase mb-2 mt-0"
-              style={{ color: "var(--txt-dim)" }}
-            >
-              Others
-            </h3>
-          )}
-          <div
-            className="grid gap-2"
-            style={{
-              gridTemplateColumns: "repeat(auto-fill, minmax(200px, 1fr))",
-            }}
-          >
-            {unpinnedNotes.map((note) => (
-              <NoteCard
-                key={note.id}
-                note={note}
-                onSelect={setSelectedNote}
-                onPin={togglePin}
-                onDelete={deleteNote}
-                onDuplicate={duplicateNote}
-                onExport={exportNote}
-                onColorChange={changeColor}
-                showColorPicker={showColorPicker}
-                setShowColorPicker={setShowColorPicker}
-                colors={colors}
-                getPlainTextPreview={getPlainTextPreview}
-              />
-            ))}
-          </div>
-        </div>
-      )}
-
-      {/* Empty state */}
-      {filteredNotes.length === 0 && (
-        <div className="text-center mt-10" style={{ color: "var(--txt-dim)" }}>
-          {searchTerm
-            ? "No notes found"
-            : "No notes yet. Create your first note!"}
-        </div>
-      )}
-    </div>
-  );
-};
-
-export default NotesList;
